Split cgroups v1 and v2
Move the implementation of CPU values for v1 and v2 cgroups to
separate files. Move the v1 and v2 cgroups test fixtures to
separate folders.
