+++ fixtures/linux/sys/fs/cgroup_v1/cpuacct_1/cpuacct.stat
