+++ fixtures/linux/sys/fs/cgroup_v1/cpuacct_2/cpuacct.stat
