+++ fixtures/linux/sys/fs/cgroup_v1/cpuacct_garbage/cpuacct.stat
