+++ fixtures/linux/sys/fs/cgroup_v1/memory/memory.stat
