+++ fixtures/linux/sys/fs/cgroup_v1/memory/memory.usage_in_bytes
