+++ fixtures/linux/sys/fs/cgroup_v1/memory_incomplete/memory.stat
