+++ fixtures/linux/sys/fs/cgroup_v1/memory_without_swap/memory.usage_in_bytes
