+++ fixtures/linux/sys/fs/cgroup_v2/cpu.stat_2
