+++ src/cpu/mod.rs
@@ -1,2 +1,4 @@
 pub mod cgroup;
+mod cgroup_v1;
+mod cgroup_v2;
 pub mod proc;
